% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFAsim.R
\docType{package}
\name{SFAsim-package}
\alias{SFAsim}
\alias{SFAsim-package}
\title{Package for simulating stochastic frontier analysis (SFA) data}
\description{
This package provides a collection of methods for doing SFA (Stochastic frontier analysis).
}
