% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{sim_data}
\alias{sim_data}
\title{generate simulated (cross-section or balanced panel) data for testing purposes}
\usage{
sim_data(k = 40, t = 10, ineff_sign = -1, inputs = c("random", "fe",
  "ar"), inputs_params = list(), frontier_coef = c(10, 0.5, -0.1),
  inefficiency = c("random", "fe", "ar"), ineff_params = list(),
  frontier_sigma = 1.5, determinants = c("none", "random", "fe", "ar"),
  deter_params = list(), deter_coef = c(0.1, 0.4), aslist = FALSE)
}
\arguments{
\item{k}{number of individuals (cross section dimension)}

\item{t}{number of observations per individual (time dimension)}

\item{ineff_sign}{production (-1) or cost inefficiency form (1),}

\item{inputs}{model for inputs data}

\item{inputs_params}{parameters for model for inputs data}

\item{frontier_coef}{coefficients for explanatory variables}

\item{inefficiency}{model for inefficiency}

\item{ineff_params}{parameters for model for inefficiency}

\item{frontier_sigma}{sqrt of variance of the symmetric error}

\item{determinants}{model for inefficiency exogeneous determinants}

\item{deter_params}{parameters for model for inefficiency exogeneous determinants}

\item{deter_coef}{coefficients on exogeneous determinants}

\item{aslist}{TRUE/FALSE to return list or array}

\item{ineff_covar_coef}{coefficients for exogeneous varaibles}

\item{ineff_sigma}{variance of the inefficiency term}
}
\description{
generate simulated (cross-section or balanced panel) data for testing purposes
}
